backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
httpx
python-dotenv
pydantic